Move query bank into subpackage
The bank is super useful, but feels out of scope for the main package
