+++ bank/bank.go
@@ -1,4 +1,4 @@
-package sparql
+package bank
 
 import (
 	"bufio"
@@ -18,17 +18,17 @@ var (
 	spaceMatcher   = regexp.MustCompile(`\s{2,}`)
 )
 
-// LoadBank takes an io.Reader, parses its input and extracts SPARQL queries
+// Load takes an io.Reader, parses its input and extracts SPARQL queries
 // and stores them in a bank. Any query must be preceded by a comment which
 // tags the query with a name.
-func LoadBank(r io.Reader) Bank {
-	bank := make(map[string]string)
+func Load(r io.Reader) (b Bank) {
+	b = make(map[string]string)
 
 	s := bufio.NewScanner(r)
 
 	var (
 		keepLineState = false
-		b             bytes.Buffer
+		buf           bytes.Buffer
 		key           string
 	)
 
@@ -38,43 +38,48 @@ func LoadBank(r io.Reader) Bank {
 		if tagMatcher.MatchString(line) {
 			keyCandidate := tagMatcher.FindStringSubmatch(line)[1]
 			if keyCandidate != key && keepLineState {
-				bank[key] = b.String()
-				b.Reset()
+				b[key] = buf.String()
+				buf.Reset()
 			}
+
 			key = keyCandidate
 			keepLineState = true
 		}
 
 		if keepLineState && !commentMatcher.MatchString(line) {
-			b.WriteString(line)
-			b.WriteString(" ") // s.Scan() strips newlines, so ensure we got a whitespace instead
-			bank[key] = b.String()
+			buf.WriteString(line)
+			buf.WriteString(" ") // s.Scan() strips newlines, so ensure we got a whitespace instead
+			b[key] = buf.String()
 		}
 	}
 
-	for k := range bank {
-		bank[k] = stripLine(bank[k])
+	for k := range b {
+		b[k] = stripLine(b[k])
 	}
-	return bank
+
+	return
 }
 
 // Prepare returns the query string given a key, and optionally a struct with
 // exported fields to be interpolated as variables into the query.
 func (b Bank) Prepare(key string, i ...interface{}) (string, error) {
-
 	if q, ok := b[key]; ok {
 		if len(i) == 0 {
 			return q, nil
 		}
+
 		t, err := template.New("query").Parse(q)
 		if err != nil {
 			return "", err
 		}
+
 		var b bytes.Buffer
 		err = t.Execute(&b, i[0])
+
 		if err != nil {
 			return "", err
 		}
+
 		return b.String(), nil
 	}
